fix(docs): prevent keyboard from overlapping chat input on mobile
Extract useMobileKeyboardOffset hook from BlockNoteEditor's inline
keyboard detection logic (VirtualKeyboard API + VisualViewport fallback).
Apply it to ChatSidebar so the composer stays above the virtual keyboard.
