+++ packages/docs/src/components/editor/BlockNoteEditor.tsx
@@ -49,6 +49,7 @@ import { useResolveUsers } from '../../hooks/useResolveUsers';
 import { useMentionUsers } from '../../hooks/useMentionUsers';
 import { useDocsAdapter } from '../../context/DocsContext';
 import { useIsTouchDevice } from '../../hooks/useIsTouchDevice';
+import { useMobileKeyboardOffset } from '../../hooks/useMobileKeyboardOffset';
 import { useEditorPreferencesStore } from '../../stores/editorPreferencesStore';
 import { ErrorBoundary } from '../common/ErrorBoundary';
 import { Mention } from './Mention';
@@ -175,14 +176,11 @@ const BlockNoteEditorInner = ({
   const getMentionMenuItems = useMentionUsers(provider ?? null);
   const hasInitialized = useRef(false);
   const [isReady, setIsReady] = useState(false);
-  const wrapperRef = useRef<HTMLDivElement>(null);
+  const wrapperRef = useMobileKeyboardOffset<HTMLDivElement>();
 
+  // Editor-specific: toggle selection class + scroll selection into view
   useEffect(() => {
     if (!isTouchDevice) return;
-
-    const setOffset = (px: number) => {
-      wrapperRef.current?.style.setProperty('--mobile-keyboard-offset', px > 0 ? `${px}px` : '0px');
-    };
 
     const updateSelectionClass = () => {
       const sel = window.getSelection();
@@ -209,74 +207,18 @@ const BlockNoteEditorInner = ({
       }
     };
 
-    // Prefer VirtualKeyboard API (Chrome/Edge 93+) — gives exact geometry, no delay
-    const vk = (navigator as any).virtualKeyboard;
-    if (vk) {
-      vk.overlaysContent = true;
-      const onGeometryChange = () => {
-        setOffset(vk.boundingRect.height);
-        scrollTimer = setTimeout(scrollSelectionIntoView, 100);
-      };
-      vk.addEventListener('geometrychange', onGeometryChange);
-      const onSelectionChange = () => {
-        updateSelectionClass();
-        scrollSelectionIntoView();
-      };
-      document.addEventListener('selectionchange', onSelectionChange);
-      return () => {
-        vk.removeEventListener('geometrychange', onGeometryChange);
-        document.removeEventListener('selectionchange', onSelectionChange);
-        clearTimeout(scrollTimer);
-      };
-    }
-
-    // Fallback: Visual Viewport API (Safari, older browsers)
-    const vp = window.visualViewport;
-    if (!vp) return;
-
-    let lastKnownKeyboardHeight = 0;
-
-    const update = () => {
-      const layoutHeight = document.documentElement.clientHeight;
-      const keyboardHeight = layoutHeight - vp.height - vp.offsetTop;
-      if (keyboardHeight > 50) lastKnownKeyboardHeight = keyboardHeight;
-      setOffset(keyboardHeight);
+    const onSelectionChange = () => {
+      updateSelectionClass();
       clearTimeout(scrollTimer);
       scrollTimer = setTimeout(scrollSelectionIntoView, 100);
     };
 
-    const onFocusIn = (e: FocusEvent) => {
-      const target = e.target as HTMLElement;
-      if (target.isContentEditable || target.tagName === 'INPUT' || target.tagName === 'TEXTAREA') {
-        if (lastKnownKeyboardHeight > 0) {
-          setOffset(lastKnownKeyboardHeight);
-        }
-      }
-    };
-
-    const onFocusOut = () => {
-      setOffset(0);
-    };
-
-    const onSelectionChange = () => {
-      updateSelectionClass();
-      scrollSelectionIntoView();
-    };
-
-    vp.addEventListener('resize', update);
-    vp.addEventListener('scroll', update);
-    document.addEventListener('focusin', onFocusIn);
-    document.addEventListener('focusout', onFocusOut);
     document.addEventListener('selectionchange', onSelectionChange);
     return () => {
-      vp.removeEventListener('resize', update);
-      vp.removeEventListener('scroll', update);
-      document.removeEventListener('focusin', onFocusIn);
-      document.removeEventListener('focusout', onFocusOut);
       document.removeEventListener('selectionchange', onSelectionChange);
       clearTimeout(scrollTimer);
     };
-  }, [isTouchDevice]);
+  }, [isTouchDevice, wrapperRef]);
 
   const collaborationUser = useMemo(() => {
     if (!provider?.awareness) return null;
